Remote: Fix DNS problem
This is a follow up of the revert in the previous patch.

Until we have the tunnel up and running we want to have our own DNS as
the first one.

After the tunnel is up we have to remove it and leave only 1, because
having 2 creates problems for OVN services.