Remote: Workaround for OVN cluster controller bug
It looks like when the OVN SB has no attachment network it doesn't
create the ovncontroller-config that is necessary for edpm nodes.

Looking at the code [1] it should be enough with the service override we
have, but somehow it isn't.

Until this is resolved we leave the networkAttachment as it is.

[1]: https://github.com/openstack-k8s-operators/ovn-operator/blob/ce1c1efacce1105a44eee1eaee82359382394e92/controllers/ovndbcluster_controller.go#L636