Remote: Fix MTU in OCP cluster
By default the cluster sets the MTU of the internal network to 1400 (100
less than the default interface).

This may be greater than what we can allow for packets not to get
dropped. So we need to change it.

There is no way to change this when starting the CRC VM, so it needs to
be changed afterwards, and the procedure takes a very long time to
complete.

This patch adds a new function called `change_cluster_mtu` to the
`env-vars` file to do this automatically.