Remote: Fix DNS broken after reboot
After we reboot the edpm node services won't be able to connect to the
control plane because we will have 2 DNS servers.

In commit b4ce0096c81c5097c868318022b36383b16f2e9f we fixed this issue
for the running cluster, but we didn't take into account that the values
in /etc/resolv.conf will be reverted on reboot.

This patch ensure we change `ifcfg-br-ex` so that there's only 1 DNS
server after reboot.