Remote: Support RH content
The repository only supported running with upstream operators and
OpenStack containers.

With this patch it is now possible to deploy using all Red Hat content:
Operators and OpenStack release.

To do so it's necessary to change in `env-vars` the variable
`USE_UPSTREAM` to `true`, and at the very least then set `SM_USERNAME`
and `SM_PASSWORD` with our subscription credentials.

This has been tested with RHOSO and RHEL 9.4 on the EDPM node.