Remote: Explain edpm deployment timeouts
Sometimes deploying the edpm node timeouts, and we don't have any
explanation on how to monitor the progress or see if it has actually
failed.

We also don't have any explanation on how to discover the edpm node if
the provisioning has timedout but then completed.

This patch updates the README.md file with some basic information.