Remote: Fix DNS
In some scenarios dnf will fail to resolve the repo because it's using
the OCP DNS, which is in the remote location, but the domain name for
the closest repo is not resolvable by the OCP cluster DNS, so it will
fail on the DNS resolution.

The original nameserver used by the host is able to resolve the domain
name, so we need to leave that nameserver until the very end instead of
removing it as soon as we have the tunnel to the OCP cluster up and
running.

This patch splits the tunnel from the DNS fixing, moving the fixing to
the end of the process, which will remove the "normal" nameserver as
well as restart all the containers to ensure they use the new
`resolv.conf`.