Remote: Fix VLANs MTUs
We were incorrectly setting the VLAN MTUs to be the same as the
underlying interface. This only works when the underlying interface is a
real NIC, because the NIC will just send a slightly larger packet which
all the other NICs in the network can support.

This is what was breaking the OVN controller and metadata agent and why
we were using it with the load balancer.

Now that we set the right MTU size for VLANs (to account for the 4 bytes
overhead) we no longer need to use the load balancer for OVN.