Add openstack_init target
Recently, a new target called openstack_init has been added to
install_yamls [1]. This target ensures deploynment of all the
openstack operators. The current state of this repository fails with the
following error:

no matches for kind "NetConfig"

when running make deploy_controlplane. This is caused by missing CRDs
in the environment because only openstack-operator is being deployed.

[1] https://github.com/openstack-k8s-operators/install_yamls/pull/963