Ensure EDPM node doesn't have IPv6 enabled.
Some times our EDPM nodes have .nmconnections that explicitly state to
assign IPv6 addresses, which is problematic because the rest of the
installation will follow this explicit requirement regardless of sysctl
parameters set on the installation.

This patch ensures that these files are removed and then we explicitly
disable IPv6 in all interfaces that are not the local interface.