Remote: Fix tunneling
On commit bdf6f83745aa274d8583c297812a70153da8cf93 we added the EXISTS
variable to the bash script in the tunneling.yaml file to remove the
error when running, but we didn't update the check accordingly, so it
would never enable the systemd unit and the tunnel would never come up.

In this patch we update the check to work as intended.